app/requirements.txt
django==3.1.*
psycopg2==2.8.*
pysocks
argh
autoflake==1.3.*
black==19.10b0
boto3==1.12.*
celery[redis]==4.4.7
kombu==4.6.11
celery-redbeat==1.0.*
django-celery-results==1.1.*
djangorestframework==3.11.*
dj-database-url==0.5.*
environs==7.1.*
isort==4.3.*
netifaces==0.10.*
paramiko==2.7.*
python-dotenv
pyyaml==5.3.*
requests==2.22.*
selenium==3.141.*
watchdog==0.9.*
python-json-logger==0.1.11
sentry-sdk==0.14.*
redis==3.3.*
django-redis==4.11.*
mypy==0.770
django-stubs==1.5.*
djangorestframework-stubs==1.1.*
django-tables2
gunicorn[gevent]==20.0.*
psycogreen==1.0.*
django-db-geventpool==3.2.2
ddtrace==0.45.*
datadog==0.39.*
jedi==0.17.2
ipython==7.10.*
django-alive==1.1.*
